requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
cloudinary==1.28.1
dj-database-url==0.5.0
Django==4.0.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.3
Pillow==9.0.0
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
